Changing file-sorting method
File-sorting used to be done alphabetically, as I thought files were
named alphabetically by date. This is GENERALLY true, but the naming
scheme changes from time to time!

Now, sorting is done by the last-modified date in Windows (NOT
date-created, as this date is actually inaccurate).

TODO: change the sketchy method by which files are obtained from a given
directory, and (hopefully) be able to select ACQ files to concatenate
from any directory.