Handling issues with BM-21, BM-22
In the ACQ files, BM-22(r) is labeled incorrectly.

BM-21(l) is correct, but BM-22(r) SHOULD be BM-21(r). This causes issues
-- BM-22 is interpreted as a different animal, and the (l) and (r)
channels are not paired for BM-21.

This has been fixed by specifically handling this typo, and removing
BM-22 from the list of animals if it exists there. There are no other
similar typos that I know of.